+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   was required.
 - Speed up many filters by caching class constants.
 - Add support for callbacks around `execute`.
+- Allow callable defaults.
 
 # [1.0.5][] (2014-02-25)